fix: minor fixes & improve docker container services health checks

+++ pkg/db/migrations/001_init.up.sql
@@ -1,14 +1,14 @@
 CREATE TABLE orders
 (
     order_uid          TEXT PRIMARY KEY,
-    track_number       TEXT        NOT NULL,
-    entry              TEXT        NOT NULL,
-    locale             TEXT        NOT NULL,
-    internal_signature TEXT        NOT NULL,
-    customer_id        TEXT        NOT NULL,
-    delivery_service   TEXT        NOT NULL,
+    track_number       TEXT NOT NULL,
+    entry              TEXT NOT NULL,
+    locale             TEXT NOT NULL,
+    internal_signature TEXT NOT NULL,
+    customer_id        TEXT NOT NULL,
+    delivery_service   TEXT NOT NULL,
     shardkey           TEXT,
-    sm_id              BIGINT      NOT NULL,
+    sm_id              BIGINT NOT NULL,
     date_created       TIMESTAMPTZ NOT NULL,
     oof_shard          TEXT
 );
@@ -28,13 +28,13 @@ CREATE TABLE deliveries
 CREATE TABLE payments
 (
     order_uid     TEXT PRIMARY KEY REFERENCES orders (order_uid) ON DELETE CASCADE,
-    transaction   TEXT   NOT NULL,
-    request_id    TEXT   NOT NULL,
-    currency      TEXT   NOT NULL,
-    provider      TEXT   NOT NULL,
+    transaction   TEXT NOT NULL,
+    request_id    TEXT NOT NULL,
+    currency      TEXT NOT NULL,
+    provider      TEXT NOT NULL,
     amount        BIGINT NOT NULL,
     payment_dt    BIGINT NOT NULL,
-    bank          TEXT   NOT NULL,
+    bank          TEXT NOT NULL,
     delivery_cost BIGINT NOT NULL,
     goods_total   BIGINT NOT NULL,
     custom_fee    BIGINT NOT NULL
@@ -42,17 +42,17 @@ CREATE TABLE payments
 
 CREATE TABLE items
 (
-    id           SERIAL PRIMARY KEY,
-    order_uid    TEXT   NOT NULL REFERENCES orders (order_uid) ON DELETE CASCADE,
+    id           BIGSERIAL PRIMARY KEY,
+    order_uid    TEXT NOT NULL REFERENCES orders (order_uid) ON DELETE CASCADE,
     chrt_id      BIGINT NOT NULL,
-    track_number TEXT   NOT NULL,
+    track_number TEXT NOT NULL,
     price        BIGINT NOT NULL,
-    rid          TEXT   NOT NULL,
-    name         TEXT   NOT NULL,
+    rid          TEXT NOT NULL,
+    name         TEXT NOT NULL,
     sale         BIGINT NOT NULL,
-    size         TEXT   NOT NULL,
+    size         TEXT NOT NULL,
     total_price  BIGINT NOT NULL,
     nm_id        BIGINT NOT NULL,
-    brand        TEXT   NOT NULL,
+    brand        TEXT NOT NULL,
     status       BIGINT NOT NULL
 );
